cleaned up some files and changed apjRoutes to apiRoutes.. moved login button into form
