+++ public/js/dashboard.js
@@ -2,7 +2,7 @@ $(document).ready(() => {
   // To Do : Add logic for dashboard
   // This file just does a GET request to figure out which user is logged in
   // and updates the HTML on the page
-  // $.get("/apj/user_data").then(data => {
+  // $.get("/api/user_data").then(data => {
   //   $(".member-name").text(data.email);
   // });
 });
