+++ public/js/login.js
@@ -25,7 +25,7 @@ $(document).ready(() => {
 
   // loginUser does a post to our "api/login" route and if successful, redirects us the the members page
   function loginUser(email, password) {
-    $.post("/apj/login", {
+    $.post("/api/login", {
       email: email,
       password: password
     })
