+++ public/js/signup.js
@@ -12,7 +12,7 @@ $(document).ready(() => {
   // Does a post to the signup route. If successful, we are redirected to the members page
   // Otherwise we log any errors
   function signUpUser(email, password) {
-    $.post("/apj/signup", {
+    $.post("/api/signup", {
       email: email,
       password: password
     })
