+++ routes/passportRouter.js
@@ -17,7 +17,7 @@ Router.post("/signup", (req, res) => {
     password: req.body.password
   })
     .then(() => {
-      res.redirect(307, "/apj/login");
+      res.redirect(307, "/api/login");
     })
     .catch(err => {
       res.status(401).json(err);
