+++ server.js
@@ -30,8 +30,8 @@ app.set("view engine", "handlebars");
 // Requiring our routes
 require("./routes/html-routes.js")(app);
 // require("./routes/api-routes.js")(app);
-const apjRoutes = require("./routes/apjRoutes");
-app.use("/apj", apjRoutes);
+const apiRoutes = require("./routes/apiRoutes");
+app.use("/api", apiRoutes);
 
 // Syncing our database and logging a message to the user upon success
 db.sequelize.sync().then(() => {
