+++ views/login.handlebars
@@ -19,13 +19,13 @@
                     <input type="password" class="validate" id="password-input" placeholder="Password">
                   </div>
                 </div>
-
+              <div class="card-action">
+                            <button type="submit" class="btn">Login</button>
+                            <p>Or sign up <a href="/">here</a></p>
+                          </div>
               </form>
             </div>
-            <div class="card-action">
-              <button type="submit" class="btn">Login</button>
-              <p>Or sign up <a href="/">here</a></p>
-            </div>
+            
 
           </div>
         </div>
